VS2015 - Fixing a host of compiler issues when using VS2015 toolset
Lots of shadow variable issues in header files
Type narrowing mismatches between integer types.
Fixed some variable names which conflicted with private members, this works around a bug in VC14

Commiting some of Jeff's code changes from 2605975

#platformnotify Josh.Adams

#CodeReview Jeff.Campeau

[CL 2613285 by Terence Burns in Main branch]

+++ Engine/Source/Runtime/Engine/Private/TriangleRendering.cpp
@@ -288,8 +288,8 @@ bool FCanvasTriangleRendererItem::Render_GameThread(const FCanvas* Canvas)
 	{
 		View,
 		Data,
-		bNeedsToSwitchVerticalAxis,
-		Canvas->IsHitTesting(),
+		(uint32)bNeedsToSwitchVerticalAxis,
+		(uint32)Canvas->IsHitTesting(),
 		Canvas->GetAllowedModes()
 	};
 	ENQUEUE_UNIQUE_RENDER_COMMAND_ONEPARAMETER(
